refs #6, adding notifications
This adds an "enable notifications" box to the beta client.
When this is enabled it simply creates a regex out of the user name
as well as a regex to make sure the user isn't the one talking.
If notifications are not available in the browser, the button to enable
should not appear at all.  If for some reason it does, when the user
tries to enable it will alert them with native Alert message.

Future work: The notifications themselves could probably be made a bit prettier.